Create a get_diff_dir() function to diff and report against a directory
Rather than diff'ing and reporting on every file in a directory created
by copy_dir, we run the diff and only report changed files, or if none
have changed then report that nothing has changed in the directory.

The PR also contains a change to the copy_dir function to include a
file_pattern option. The copy_dir function will only copy file matching
the regex pattern. Useful for the network-scripts directory and copying
ifcfg-* and route-* files for example.